added install, clear  and updated init

+++ PowerManager/Public/Clear-PowerManager.ps1
@@ -4,9 +4,11 @@ function Clear-PowerManager {
     This will clear out the PowerManager cache.  Use this if you are experiencing odd behaviour with your modules.
     #>
     param (
-
+        [switch] $Modules
     )
-    $Path = "$PWD/.pm/modules"
-    Remove-Item -Path $Path -Force
-    $r = New-Item -Path $Path
+    if ($Modules) {
+        $Path = "$PWD/.pm/modules"
+        Remove-Item -Path $Path -Force
+        $r = New-Item -Path $Path
+    }
 }
